+++ CHANGELOG.md
@@ -156,6 +156,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - Updated `data-migration1` and `data-migration2` lambdas to return `MigrationSummary` objects.
     - Added logging for every batch of 100 records processed for executions, granules and files, and PDRs.
     - Remove `RecordAlreadyMigrated` logs in `data-migration1` and `data-migration2`
+  - **CUMULUS-2452**
+    - Added support for only migrating certain granules by specifying the `granuleSearchParams.granuleId` or `granuleSearchParams.collectionId` properties in the payload for the `<prefix>-postgres-migration-async-operation` Lambda
+    - Added support for only running certain migrations for data-migration2 by specifying the `migrationsList` property in the payload for the
+    `<prefix>-postgres-migration-async-operation` Lambda
 
 ## [v7.1.0] 2021-03-12